requirements.txt
python-dotenv
requests
hmac
urllib
hashlib